Refactor: 패키지 구조 변경
StationInfo 클래스 위치 이동
