+++ src/main/java/podo/odeego/domain/place/service/PlaceQueryService.java
@@ -2,7 +2,7 @@ package podo.odeego.domain.place.service;
 
 import podo.odeego.domain.place.domain.PlaceCategory;
 import podo.odeego.domain.place.dto.PlaceResponses;
-import podo.odeego.domain.subway.dto.StationInfo;
+import podo.odeego.domain.station.dto.StationInfo;
 
 public interface PlaceQueryService {
